requirements.txt
pandas
aiohttp
python-dotenv